[Minghao] #39 add 'category' field in response

+++ src/main/java/org/siglus/siglusapi/dto/android/response/ProductChildResponse.java
@@ -18,8 +18,8 @@ package org.siglus.siglusapi.dto.android.response;
 import java.util.Map;
 import java.util.UUID;
 import lombok.Data;
-import org.openlmis.stockmanagement.dto.referencedata.OrderableChildDto;
-import org.openlmis.stockmanagement.dto.referencedata.OrderableDto;
+import org.siglus.common.dto.referencedata.OrderableChildDto;
+import org.siglus.common.dto.referencedata.OrderableDto;
 
 @Data
 public class ProductChildResponse {
@@ -31,7 +31,7 @@ public class ProductChildResponse {
       Map<UUID, OrderableDto> productMap) {
     ProductChildResponse resp = new ProductChildResponse();
     resp.setProductCode(productMap.get(child.getOrderable().getId()).getProductCode());
-    resp.setQuantity(child.getQuantity().longValue());
+    resp.setQuantity(child.getQuantity());
     return resp;
   }
 }
